Add conditions for deleting new tempfile in `MiniMagick::Image#format`.
Add conditions for deleting new tempfile if it exist and it is not the same as current tempfile when convert fails in `MiniMagick::Image#format`.

+++ lib/mini_magick/image.rb
@@ -462,7 +462,7 @@ module MiniMagick
 
       self
     rescue MiniMagick::Invalid, MiniMagick::Error => e
-      new_tempfile.unlink
+      new_tempfile.unlink if new_tempfile && @tempfile != new_tempfile
       raise e
     end
